added all logic
