+++ src/components/OrderDetails/OrderDetails.jsx
@@ -9,18 +9,30 @@ import React from "react";
 import doneIndicationImg from "../../images/done.png";
 import stylesOrderDetails from "../OrderDetails/OrderDetails.module.css";
 import { useSelector } from "react-redux";
+import stylesIngredientDetails from "../IngredientDetails/IngredientDetails.module.css";
+import {InfinitySpin} from "react-loader-spinner";
 
 const OrderDetails = () => {
-    const orderNumber = useSelector((store) => store.order.order);
+    const order = useSelector((state) => state.order);
+    console.log(order)
+
 
     return (
-        <section className={`${stylesOrderDetails.container} pb-20`}>
-            <p className={`${stylesOrderDetails.title} text text_type_digits-large`}>{orderNumber === 0 ? `loading...` : orderNumber}</p>
+        <>
+        { order ? (
+        <div className={`${stylesOrderDetails.container} pb-20`}>
+            <p className={`${stylesOrderDetails.title} text text_type_digits-large`}>{order?.order === 0 ? `loading...` : order.order}</p>
             <p className={`text text_type_main-medium pt-8 pb-15`}>идентификатор заказа</p>
             <img className={stylesOrderDetails.image} src={doneIndicationImg} alt={"finished"} />
             <p className={`text text_type_main-default pt-15 pb-2`}>Ваш заказ начали готовить</p>
             <p className={`text text_type_main-default text_color_inactive`}>Дождитесь готовности на орбитальной станции</p>
-        </section>
+        </div>
+            ) : (
+                <div className={stylesIngredientDetails.loader}>
+                    <InfinitySpin width="200" color="#4c4cff" />
+                </div>
+            )}
+        </>
     );
 };
