+++ src/components/ProtectedRoute/ProtectedRoute.jsx
@@ -6,7 +6,7 @@ import PropTypes from "prop-types";
 export const ProtectedRoute = ({ onlyForAuth, children, ...rest }) => {
     const isAuthorized = getCookie("accessToken");
     const isAuth = localStorage.getItem('refreshToken')
-    console.log(isAuth)
+
 
     const location = useLocation();
 
@@ -26,6 +26,7 @@ export const ProtectedRoute = ({ onlyForAuth, children, ...rest }) => {
             </Route>
         );
     }
+
 
     return <Route {...rest}>{children}</Route>;
 };
