+++ src/pages/OrderPage/OrderPage.jsx
@@ -21,6 +21,8 @@ export const OrderPage = () => {
     const location = useLocation();
 
 
+
+
     useEffect( () => {
         location.pathname.indexOf('/profile/orders') !== -1
             ?
