+++ src/services/actions/order.jsx
@@ -6,7 +6,8 @@ export const SEND_ORDER_SUCCESS = "SEND_ORDER_SUCCESS";
 export const SEND_ORDER_FAILED = "SEND_ORDER_FAILED";
 export const GET_ORDER = 'GET_ORDER';
 export const GET_ORDER_SUCCESS = 'GET_ORDER_SUCCESS';
-export const GET_ORDER_FAILED = 'GET_ORDER_FAILED'
+export const GET_ORDER_FAILED = 'GET_ORDER_FAILED';
+export const CLEAR_ORDER = 'CLEAR_ORDER';
 
 export const setOrder = (ingredients) => {
     return (dispatch) => {
