+++ src/services/reducers/getOrderReducer.jsx
@@ -1,5 +1,5 @@
 import {SEND_ORDER, SEND_ORDER_FAILED, SEND_ORDER_SUCCESS} from "../actions/order";
-import { GET_ORDER, GET_ORDER_FAILED, GET_ORDER_SUCCESS} from "../actions/order";
+import { GET_ORDER, GET_ORDER_FAILED, GET_ORDER_SUCCESS, CLEAR_ORDER} from "../actions/order";
 
 const initialState = {
     price: 0,
@@ -9,7 +9,8 @@ const initialState = {
     user: {},
     order: {},
     orderList: [],
-    isLoad: false
+    isLoad: false,
+
 };
 
 export const getOrderData = (state = initialState, action) => {
@@ -38,7 +39,7 @@ export const getOrderData = (state = initialState, action) => {
         // case CLEAR_ORDER: {
         //     return {
         //         ...state,
-        //         order: 0
+        //         ...action.payload
         //     }
         // }
         default: {
